Separate subscriptions logic into a custom hook and a separate component
